pyproject.toml
[project]
name = "dslrecommendationsystem"
version = "0.1.0"
description = "DSL recommendation system for the JetBrains MPS ecosystem"
authors = [
    {name = "Ana Carolina Poltronieri Rodrigues",email = "anapoltronieri.aluno@unipampa.edu.br"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "google-genai (>=1.68.0,<2.0.0)",
    "fastapi (>=0.136.1,<0.137.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.0,<27.0.0)",
    "ruff (>=0.15.5,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]
[tool.poetry]
package-mode = false